added about page hero images
